fix: addressed comments and fixed tests
Fixed comments

RESOLVES: OKTA-301556
