+++ packages/@okta/i18n/src/properties/login.properties
@@ -703,13 +703,16 @@ cert.authentication.title = Certificate authentication
 success.redirect = You will be redirected
 
 ##-- OIE strings
+
+oie.phone = Phone
+
 # OIE Email
 oie.email.mfa.title = Verify with your email
 
 #OIE webauthn
 oie.webauthn = Security Key or Biometric Authenticator
 oie.webauthn.description = Use a security key or a biometric authenticator to sign in
-oie.enroll.webauthn.title=Set up security key or biometric authenticator
+oie.enroll.webauthn.title= Set up security key or biometric authenticator
 oie.enroll.webauthn.instructions = Your browser will prompt to register a security key or biometric authenticator (Windows Hello, Touch ID, etc.). Follow the instructions to complete enrollment.
 oie.enroll.webauthn.uv.required.instructions = Biometric verification or a PIN is required to setup this authenticator.
 oie.verify.webauthn.instructions = Your browser will prompt to register a security key or biometric authenticator (Windows Hello, Touch ID, etc.). Follow the instructions to complete enrollment.
