+++ src/v2/util/FactorUtil.js
@@ -25,7 +25,7 @@ const factorData = {
   },
 
   'phone': {
-    label: loc('factor.phone', 'login'),
+    label: loc('oie.phone', 'login'),
     description: '',
     // TODO: need new icon for `mfa-okta-phone`
     iconClassName: 'mfa-okta-call',
