+++ test/testcafe/spec/SelectAuthenticatorForVerification_spec.js
@@ -40,15 +40,15 @@ test.requestHooks(mockChallengePassword)(`should load select authenticator list`
   await t.expect(selectFactorPage.getFormSubtitle()).eql('Verify with one of the following factors.');
   await t.expect(selectFactorPage.getFactorsCount()).eql(3);
 
-  await t.expect(selectFactorPage.getFactorLabelByIndex(0)).eql('Okta Password');
+  await t.expect(selectFactorPage.getFactorLabelByIndex(0)).eql('Password');
   await t.expect(selectFactorPage.getFactorIconClassByIndex(0)).contains('mfa-okta-password');
   await t.expect(selectFactorPage.getFactorSelectButtonByIndex(0)).eql('Select');
 
-  await t.expect(selectFactorPage.getFactorLabelByIndex(1)).eql('Security Key or Biometric Authenticator (FIDO2)');
+  await t.expect(selectFactorPage.getFactorLabelByIndex(1)).eql('Security Key or Biometric Authenticator');
   await t.expect(selectFactorPage.getFactorIconClassByIndex(1)).contains('mfa-webauthn');
   await t.expect(selectFactorPage.getFactorSelectButtonByIndex(1)).eql('Select');
 
-  await t.expect(selectFactorPage.getFactorLabelByIndex(2)).eql('Okta Email');
+  await t.expect(selectFactorPage.getFactorLabelByIndex(2)).eql('Email Authentication');
   await t.expect(selectFactorPage.getFactorIconClassByIndex(2)).contains('mfa-okta-email');
   await t.expect(selectFactorPage.getFactorSelectButtonByIndex(2)).eql('Select');
